chore(veritech-client): extract unit tests as integration tests
Guess what? `lib/veritech-client`'s "unit" tests were in fact,
integration tests all along!! As the author of these tests waaaay back
it took until working on the new build and ci/cd systems (that being
Buck2 and Buildkite) that this mismatch revealed itself.

We would generally consider unit tests to be fairly standalone, from the
perspective of required supporting services (think databases, message
queues, etc.). As Veritech client's tests are directly dependent on a
running Veritech server (running as a library per-test) which relies on
NATS as the messaging layer, it seems more reasonable to consider these
integration tests.

Note that our upcoming CI/CD system doesn't provide pg/nats/otelcol
servies to unit tests, but does to integration tests. The failure of
`lib/veritech-client`'s unit tests in CI were the hint here. Kinda cool.
